+++ simplex.h
@@ -11,6 +11,6 @@ void update_prnt(Arvore, Arc); /*atualiza o vetor de pais e de arcos da arvore*/
 void update_depth(Arvore); /*atualiza o vetor de profundidades*/
 void update_y(Arvore); /*atualiza o vetor de potenciais*/
 void network_simplex(Graph, Arvore); /*Roda o simplex de redes, tendo como entrada um Graph - a configuracao da rede e uma Arvore, com uma solucao viavel*/
-float net_cost(Arvore, Graph); /*Devolve o custo de transporte dada uma Arvore contendo uma solucao*/
+double net_cost(Arvore, Graph); /*Devolve o custo de transporte dada uma Arvore contendo uma solucao*/
 
 #endif
